+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HEAD
 
 - Adding: attr_reader for `:path, :html_options, :width, :height`
+- Deprecating: the use of `:size` for "large" or a number, instead use `:height`
 
 # 1.1.0